adapt map_harmonically_forced and move to calculator.py
enable it to be used for both harmonic optimizations (adsorbate guess and TS guess)
enable it to be run in parallel with joblib
